Backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
chromadb==0.4.22
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langgraph==0.2.14
langsmith==0.1.98
sentence-transformers==2.3.1
pypdf2==3.0.1
python-docx==1.1.0
pillow==10.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
groq==0.4.1
python-dotenv==1.0.0
pandas==2.2.0
openpyxl==3.1.2
jinja2==3.1.3